+++ tests/e2e/cases/dynamic_require_manifest_affix/plugins/current.plugin.ts
@@ -1,0 +1,2 @@
+console.log("plugin current init");
+module.exports.label = "plugin-current";